+++ sym/tests/aes_cbc_enc_tester.c
@@ -156,7 +156,7 @@ LC_TEST_FUNC(int, main, int argc, char *argv[])
 	(void)argv;
 
 	if (argc >= 2)
-		lc_init(LC_INIT_AES_CT);
+		lc_init(LC_INIT_AES_SBOX);
 
 	LC_EXEC_ONE_TEST(lc_aes_cbc);
 	LC_EXEC_ONE_TEST(lc_aes_cbc_aesni);