+++ CHANGES.md
@@ -1,5 +1,7 @@
 Changes 1.7.1-prerelease
-* Offer a means to select the AES-C constant time implementation via lc_init API
+* Offer a means to select the AES-C constant time / S-Box implementation via lc_init API
+
+* use the AES-C constant time implementation by default - it is about 3 times slower than the AES-C S-Box implementation, but more secure. As the leancrypto library is about secure by default, the CT implementation is just right. Furthermore, if a caller wants to have the faster AES-C S-Box, he can call lc_init(LC_INIT_AES_SBOX) at the beginning.
 
 Changes 1.7.0
 * FIPS: mark only seeded DRBG instances as FIPS-approved